Hotfix testing profile

+++ practice_1/ecommerce/src/main/kotlin/com/api/ecommerce/domains/Category.kt
@@ -14,7 +14,7 @@ data class Category(
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    val id: Int = -1) {
+    val id: Long = -1) {
 
     constructor() : this("", "", -1)
 }
